dev/requirements.txt
autogluon>=0.8.0
jupyter>=1.0.0
numpy>=1.21.0
openai>=1.0.0
pandas>=1.3.0
python-dotenv>=1.0.0
pyyaml>=6.0
requests>=2.28.0
scikit-learn>=1.0.0
seaborn>=0.11.0
tqdm>=4.62.0
openpyxl>=3.1.5